backend/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.0

SQLAlchemy==2.0.35
psycopg[binary]==3.2.13 

python-dotenv==1.0.1
gunicorn==22.0.0
